+++ BotServer/package.json
@@ -18,12 +18,12 @@
   "author": "Frank Vanris",
   "license": "ISC",
   "devDependencies": {
-    "@babel/core": "7.22.10",
+    "@babel/core": "^7.22.10",
     "@babel/preset-env": "^7.26.0",
-    "@babel/preset-react": "^7.25.9",
+    "@babel/preset-react": "^7.26.3",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.0.1",
-    "babel-loader": "^9.1.3",
+    "babel-loader": "^9.2.1",
     "babel-plugin-styled-components": "2.1.4",
     "browser-sync-webpack-plugin": "^2.3.0",
     "copy-webpack-plugin": "^12.0.2",
