+++ BotServer/src/components/NavBar.jsx
@@ -1,7 +1,7 @@
 //import { PriceChangeOutlined } from "@mui/icons-material"
 import React from 'react';
 import { Link, useMatch,useResolvedPath } from "react-router-dom"
-import AccountMenu from './Login';
+import AccountMenu from './AccountMenu.jsx';
 import {Header, LinkItem} from '../styles/NavBar.style';
 
 import Logo from '../res/logo.png';
